+++ app/lib/Wanez/GrimDawn/wzMath.js
@@ -21,7 +21,9 @@ module.exports = {
             mul: false,
             incEvery: 1,
             incInc: 0,
-            incOnceEvery: 1
+            incOnceEvery: 1,
+            parabola: 0,
+            parabolaStr: 0
         },$opt || {});
         
         let $number = parseFloat($opt.number),
@@ -30,12 +32,14 @@ module.exports = {
             $incInc = parseFloat($opt.incInc),
             $max = $opt.max,
             prevInc = 0,
+            parabola = parseFloat($opt.parabola),
+            parabolaStr = parseFloat($opt.parabolaStr) || 0.2,
             tempInc = 0;
         
         let aValues = [];
         
         for(let i=1; i<=$max; i++){
-            
+    
             if($opt.round === `up`){
                 //$start = Math.ceil($start);
                 aValues.push(Math.ceil($start).toFixed($opt.dec));
@@ -47,9 +51,19 @@ module.exports = {
             }
             
             tempInc = (i / $opt.incEvery) % 1;
+    
             if(tempInc === 0){
                 //prevInc = tempInc;
                 $number = $number + $incInc;
+            }
+            if(parabola > 0){
+                let x = i / parabola;
+                Log(x);
+                if(x > 1){
+                    $number = (x - parabolaStr) * parseFloat($opt.number);
+                }else{
+                    $number = (1 + (parabolaStr - x)) * parseFloat($opt.number);
+                }
             }
             
             if((i / $opt.incOnceEvery) % 1 === 0){
@@ -67,6 +81,9 @@ module.exports = {
             if($numberMax > 0 && $start > $numberMax){
                 $start = $numberMax;
             }
+            
+    
+            
         }
         
         return aValues;